Move core test classes
Move test classes from junit package in main source set
into etomica package in test source set
